services/identity/requirements.txt
fastapi==0.115.5
mangum==0.19.0
pydantic==2.10.3
pydantic-settings==2.6.1
pyjwt[crypto]==2.10.1
bcrypt==4.2.1
pyotp==2.9.0
boto3==1.35.76
cryptography==44.0.0
uvicorn[standard]==0.32.1  # local dev only
